reduce fuzzing range on universal balance tests to match the total supply restrictions of listed tokens

+++ tests/architecture/NativeUniversalBalance/functions/DepositFor.t.sol
@@ -44,7 +44,7 @@ contract NativeUniversalBalanceDepositForTest is
     function test_nativeUniversalBalanceDepositFor_fail_whenInsufficientBalance_fuzzed(
         uint256 amount
     ) public {
-        vm.assume(amount < type(uint256).max);
+        vm.assume(amount < type(uint216).max);
 
         _prepareWETH(user1, amount);
 
@@ -61,7 +61,7 @@ contract NativeUniversalBalanceDepositForTest is
     function test_nativeUniversalBalanceDepositFor_fail_whenExceedsAllowance_fuzzed(
         uint256 amount
     ) public {
-        vm.assume(amount < type(uint256).max);
+        vm.assume(amount < type(uint216).max);
 
         _prepareWETH(user1, amount + 1);
 
@@ -115,7 +115,7 @@ contract NativeUniversalBalanceDepositForTest is
     function test_nativeUniversalBalanceDepositFor_success_withLend_fuzzed(
         uint256 amount
     ) public {
-        vm.assume(0 < amount && amount < type(uint256).max / _ONE);
+        vm.assume(0 < amount && amount < type(uint216).max / _ONE);
 
         _prepareWETH(user1, amount);
 
@@ -153,7 +153,7 @@ contract NativeUniversalBalanceDepositForTest is
     function test_nativeUniversalBalanceDepositFor_success_withoutLend_fuzzed(
         uint256 amount
     ) public {
-        vm.assume(0 < amount && amount < type(uint256).max / _ONE);
+        vm.assume(0 < amount && amount < type(uint216).max / _ONE);
 
         _prepareWETH(user1, amount);
