+++ tests/architecture/NativeUniversalBalance/functions/WithdrawFor.t.sol
@@ -73,7 +73,7 @@ contract NativeUniversalBalanceWithdrawForTest is
     function test_nativeUniversalBalanceWithdrawFor_fail_whenExceedsLentBalance_fuzzed(
         uint256 amount
     ) public {
-        vm.assume(0 < amount && amount < type(uint256).max / _ONE);
+        vm.assume(0 < amount && amount < type(uint216).max / _ONE);
 
         deal(user1, amount);
 
@@ -89,7 +89,7 @@ contract NativeUniversalBalanceWithdrawForTest is
     function test_nativeUniversalBalanceWithdrawFor_fail_whenExceedsSittingBalance_fuzzed(
         uint256 amount
     ) public {
-        vm.assume(0 < amount && amount < type(uint256).max / _ONE);
+        vm.assume(0 < amount && amount < type(uint216).max / _ONE);
 
         deal(user1, amount);
 
@@ -117,7 +117,7 @@ contract NativeUniversalBalanceWithdrawForTest is
         uint256 withdrawAmount
     ) public {
         vm.assume(
-            0 < depositAmount && depositAmount < type(uint256).max / _ONE
+            0 < depositAmount && depositAmount < type(uint216).max / _ONE
         );
         vm.assume(0 < withdrawAmount && withdrawAmount <= depositAmount);
 
@@ -163,7 +163,7 @@ contract NativeUniversalBalanceWithdrawForTest is
         uint256 withdrawAmount
     ) public {
         vm.assume(
-            0 < depositAmount && depositAmount < type(uint256).max / _ONE
+            0 < depositAmount && depositAmount < type(uint216).max / _ONE
         );
         vm.assume(0 < withdrawAmount && withdrawAmount <= depositAmount);
