+++ tests/architecture/NativeUniversalBalance/functions/WithdrawNativeFor.t.sol
@@ -42,7 +42,7 @@ contract WithdrawNativeForTest is TestBaseNativeUniversalBalance {
     function test_withdrawNativeFor_fail_whenExceedsLentBalance_fuzzed(
         uint256 amount
     ) public {
-        vm.assume(0 < amount && amount < type(uint256).max / _ONE);
+        vm.assume(0 < amount && amount < type(uint216).max / _ONE);
 
         deal(user1, amount);
 
@@ -63,7 +63,7 @@ contract WithdrawNativeForTest is TestBaseNativeUniversalBalance {
     function test_withdrawNativeFor_fail_whenExceedsSittingBalance_fuzzed(
         uint256 amount
     ) public {
-        vm.assume(0 < amount && amount < type(uint256).max / _ONE);
+        vm.assume(0 < amount && amount < type(uint216).max / _ONE);
 
         deal(user1, amount);
 
@@ -94,7 +94,7 @@ contract WithdrawNativeForTest is TestBaseNativeUniversalBalance {
         uint256 withdrawAmount
     ) public {
         vm.assume(
-            0 < depositAmount && depositAmount < type(uint256).max / _ONE
+            0 < depositAmount && depositAmount < type(uint216).max / _ONE
         );
         vm.assume(0 < withdrawAmount && withdrawAmount <= depositAmount);
 
@@ -145,7 +145,7 @@ contract WithdrawNativeForTest is TestBaseNativeUniversalBalance {
         uint256 withdrawAmount
     ) public {
         vm.assume(
-            0 < depositAmount && depositAmount < type(uint256).max / _ONE
+            0 < depositAmount && depositAmount < type(uint216).max / _ONE
         );
         vm.assume(0 < withdrawAmount && withdrawAmount <= depositAmount);
